rename downloadResults to parseResults
